FeedbackClient/requirements.txt
PyQt5
PyQt5-stubs
python-socketio[client]
requests
injector
Events
apscheduler